+++ RELEASE_NOTES.md
@@ -1,2 +1,5 @@
 - Web App Update OTA
 - Update Modal Styles
+- Backport Warning
+- Better update success response
+- Better redirect handling